requirements.txt
asgiref==3.7.2
astroid==3.1.0
colorama==0.4.6
coverage==7.4.4
dill==0.3.8
Django==5.0.3
djangorestframework==3.15.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==8.1.1
sqlparse==0.4.4
tomlkit==0.12.4
tzdata==2024.1
